Fix stuff.

+++ runtime/src/lib.rs
@@ -39,11 +39,11 @@ pub(crate) fn as_cairo_short_string(value: &Felt) -> Option<String> {
 pub unsafe extern "C" fn cairo_native__libfunc__debug__print(
     target_fd: i32,
     data: *const [u8; 32],
-    len: usize,
+    len: u32,
 ) -> i32 {
     let mut target = File::from_raw_fd(target_fd);
 
-    for i in 0..len {
+    for i in 0..len as usize {
         let mut data = *data.add(i);
         data.reverse();
